refactor: extract three wheel game hook
